--- src/main/java/com/meoguri/linkocean/controller/controller/dto/NotificationResponse.java
@@ -1,4 +1,0 @@
-package com.meoguri.linkocean.controller.controller.dto;
-
-public class NotificationResponse {
-}
