+++ src/main/java/com/meoguri/linkocean/controller/notification/dto/NotificationResponse.java
@@ -1,0 +1,4 @@
+package com.meoguri.linkocean.controller.notification.dto;
+
+public class NotificationResponse {
+}
